+++ src/components/page-components/index/Main.module.less
@@ -1,0 +1,6 @@
+.background {
+  padding: 25px 25px 25px 25px;
+  width: 1000px;
+  height: 500px;
+  background-color: #e2e2e2;
+}
